Updating azure

+++ inc/Forms/editMapForm.php
@@ -59,8 +59,9 @@ if (isset($_SESSION['username'])) {
             <p><input type='submit' value='Submit'></p>
             <?}?>
         </form>
-        <!-- Update you silly thing -->
-        <a href='/../deleteMapForm/<?php Print($locationID);?>' class="button">Delete Marker</a>
+        <? if($_SESSION['accessLevel']==11||$_SESSION['accessLevel']==31){?>
+            <a href='/../deleteMapForm/<?php Print($locationID);?>' class="button">Delete Marker</a>
+        <?}?>
         </body>
         <?
         include("/../scripts/footer.php");
